FE: Move LabelPill into taskcard folder
